+++ examples/data-sources/nxos_ospf/data-source.tf
@@ -1,0 +1,2 @@
+data "nxos_ospf" "example" {
+}