+++ examples/data-sources/nxos_ospf_instance/data-source.tf
@@ -1,0 +1,3 @@
+data "nxos_ospf_instance" "example" {
+  name = "OSPF1"
+}