+++ examples/data-sources/nxos_ospf_vrf/data-source.tf
@@ -1,0 +1,4 @@
+data "nxos_ospf_vrf" "example" {
+  instance_name = "OSPF1"
+  name          = "VRF1"
+}