+++ examples/resources/nxos_ipv4_interface_address/resource.tf
@@ -1,5 +1,5 @@
 resource "nxos_ipv4_interface_address" "example" {
   vrf          = "default"
   interface_id = "eth1/59"
-  address      = "1.1.1.1/24"
+  address      = "24.63.46.49/30"
 }