+++ examples/resources/nxos_ospf/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_ospf.example "sys/ospf"