+++ examples/resources/nxos_ospf/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_ospf" "example" {
+  admin_state = "enabled"
+}