+++ examples/resources/nxos_ospf_instance/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_ospf_instance.example "sys/ospf/inst-[OSPF1]"