+++ examples/resources/nxos_ospf_instance/resource.tf
@@ -1,0 +1,4 @@
+resource "nxos_ospf_instance" "example" {
+  admin_state = "enabled"
+  name        = "OSPF1"
+}