+++ internal/provider/data_source_nxos_subinterface_vrf_test.go
@@ -17,7 +17,7 @@ func TestAccDataSourceNxosSubinterfaceVRF(t *testing.T) {
 				Config: testAccNxosPhysicalInterfaceConfig_all(),
 			},
 			{
-				Config: testAccNxosSubinterfaceConfig_all(),
+				Config: testAccNxosPhysicalInterfaceConfig_all()+testAccNxosSubinterfaceConfig_all(),
 			},
 			{
 				Config: testAccNxosPhysicalInterfaceConfig_all()+testAccNxosSubinterfaceConfig_all()+testAccNxosSubinterfaceVRFConfig_all(),