+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Add nxos_ethernet_settings resource and data source
 - Add nxos_ospf resource and data source
 - Add nxos_ospf_instance resource and data source
+- Add nxos_ospf_vrf resource and data source
 
 ## 0.2.3